% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\title{Pairwise Linkage Disequilibrium of Ancestry (LDA)}
\usage{
LDA(data, printProcess = TRUE, SNPlimit = NA)
}
\arguments{
\item{data}{a list of data frames of the painting for ancestries.}

\item{printProcess}{logical. Print the process of calculating the pairwise LDA for the i-th SNP.}

\item{SNPlimit}{maximum number of SNPs at each side of the SNP used to calculate the pairwise LDA. The value shouldn't be larger than the total number of SNPs.}
}
\value{
a dataframe of the correlation matrix.
}
\description{
Pairwise Linkage Disequilibrium of Ancestry (LDA)
}
\examples{
library(LDAandLDAS)
#combine the painting data for two ancestries
data=list(painting_p1[,-1],painting_p2[,-1])
#calculate the pairwise LDA of SNPs
LDA_result <- LDA(data)

}
