% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\title{Pairwise Linkage Disequilibrium of Ancestry (LDA)}
\usage{
LDA(data, ancestry, printProcess = TRUE, SNPlimit = NA)
}
\arguments{
\item{data}{data frame of the painting for all the ancestries (combined in a single file).}

\item{ancestry}{number of ancestries.}

\item{printProcess}{logical. Print the process of calculating the pairwise LDA for the i-th SNP.}

\item{SNPlimit}{maximum number of SNPs at each side of the SNP used to calculate the pairwise LDA. The value shouldn't be larger than the total number of SNPs.}
}
\value{
a dataframe of the correlation matrix.
}
\description{
Pairwise Linkage Disequilibrium of Ancestry (LDA)
}
\examples{
library(LDAandLDAS)
#painting data for two ancestries
data1 <- read.csv("https://raw.githubusercontent.com/YaolingYang/LDAandLDAscore/master/painting_p1.csv")
data2 <- read.csv("https://raw.githubusercontent.com/YaolingYang/LDAandLDAscore/master/painting_p2.csv")
data=cbind(data1[,-1],data2[,-1])
#calculate the pairwise LDA of SNPs
LDA_result <- LDA(data,ancestry=2)

}
