% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAS.R
\name{LDAS}
\alias{LDAS}
\title{Linkage Disequilibrium of Ancestry Score (LDAS)}
\usage{
LDAS(LDA, map, window, printProcess = TRUE)
}
\arguments{
\item{LDA}{data frame of the pairwise LDA of all the SNPs.}

\item{map}{data frame of the physical position and genetic distance of all the SNPs.}

\item{window}{number of the window of genetic distance. The default window is 5.}

\item{printProcess}{logical. Print the process of calculating the LDA score for the i-th SNP.}
}
\value{
a data frame of the LDA score and its upper and lower bound at the physical position of each SNP.
}
\description{
Linkage Disequilibrium of Ancestry Score (LDAS)
}
\examples{
library(LDAandLDAS)
#combine the painting data for two ancestries
data=list(painting_p1[,-1],painting_p2[,-1])
#calculate the pairwise LDA of SNPs
LDA_result <- LDA(data)
#map is the data containing the physical position and recombination distance of the SNPs
#calculate the LDA score for the SNPs
LDA_score <- LDAS(LDA_result,map,window=4)

}
